requirements.txt
# Essential files for Dual Energy Source System
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
tensorflow>=2.8.0
scikit-learn>=1.0.0
flask>=2.0.0
flask-socketio>=5.0.0
pyyaml>=6.0
sqlite3

# Hardware libraries (install on Raspberry Pi)
# RPi.GPIO>=0.7.0
# adafruit-circuitpython-ads1x15>=2.2.0
# w1thermsensor>=2.0.0
# board>=1.0
# busio>=1.0

# Development and testing
pytest>=6.0.0
black>=22.0.0
pylint>=2.12.0

# Optional for advanced features
matplotlib>=3.5.0
plotly>=5.0.0
dash>=2.0.0
